feat(cli): bundle server.js + `slopstopper serve` (last template lift)
server.js was the only adopter-facing template `install.sh` still seeded into `.ss/`. The bash workflows referenced repo-root `node server.js`, which was a pre-existing inconsistency: the installer was copying to `.ss/server.js` while the workflows looked at the repo root. This PR resolves the inconsistency by lifting both copies into one canonical server inside the wheel.

- New `cli/slopstopper/data/server.js` merges the two pre-lift variants:
  - The portable auto-detect from `.ss/server.js` (probes `dist/client`, `dist`, `build`, `out`, `public`, `app`).
  - The headers-aware logic from slopstopper.dev's repo-root `server.js` (loads `worker/headers.json` when present and applies per-path CSP/COOP/CORP rules).
  - Drops the slopstopper.dev-specific `/feedback` → `/feedback.html` 301 — the pretty-URL fallback in `resolveCandidate` already serves the same HTML directly.
- `templates.py` registers `server.js` in the TEMPLATES inventory, so the override-then-bundled resolver and `slopstopper templates {list, path, eject}` all see it.
- New `slopstopper serve` subcommand: resolves `templates.template_path("server.js")` and replaces the current process with `node <path>` via `os.execvp`. The exec (not subprocess) matters for the workflow pattern `slopstopper serve &; echo $! >> $GITHUB_ENV` — bash's `$!` gives the right PID for the later `kill`, because slopstopper-cli is replaced by node in-place.
- 6 workflows flip from `node server.js &` to `slopstopper serve &` (smoke, accessibility, broken-links, SEO, CWV, DAST). The "no server.js? bail" guard goes away in each — the bundled server is always available now.
- `install.sh` stops seeding `.ss/server.js` entirely; the byte-equal scrub picks up `server.js` alongside the other lifted templates.
- Dogfood: this repo's own `server.js` and `.ss/server.js` deleted. slopstopper.dev's CI now runs the same `slopstopper serve` adopters do.
- 4 new tests: serve exec path (with stubbed `os.execvp` + `shutil.which`), `.ss/server.js` override resolution, node-missing exit code. Total: 457 passing.
- docs/{app,architecture,deployment}/README.md updated to point at `slopstopper serve` instead of the deleted file.

Net: -1 file in the adopter repo's default state. `.ss/` now contains only `reports/` and `.workflows-installed` — every other slopstopper file lives in the wheel.
