+++ .github/workflows/ss-security-dast-check.yml
@@ -58,7 +58,7 @@ jobs:
       - name: Start local server
         run: |
           npm run build
-          node server.js &
+          slopstopper serve &
           echo "SERVER_PID=$!" >> $GITHUB_ENV
           for i in $(seq 1 10); do
             if curl -s http://localhost:8080 > /dev/null; then
